+++ CHANGES.md
@@ -5,6 +5,7 @@ Change Log
 
 #### next release (mobx-36)
 * Added `pointer-events` to `MapNavigation` and `MenuBar` elements, so the bar don't block mouse click outside of the button.
+* Fixes "reminder pop-up" for help button being unclickable
 * Use `useTranslation` instead of `withTranslation` in functional component (`MapDataCount`)
 * Make welcome video url and placeholder configurable via configparameters
 * Added `ExportableData` interface.
@@ -14,6 +15,7 @@ Change Log
 * Extend `UserDrawing` to handle rectangles
 * Tsxify `MapInteractionMode`
 * Changed default orientation for `GltfCatalogItem` to no rotation, instead of zero rotation wrt to terrain
+* Added a title to welcome message video
 * (💫The next rad feature💫 but please be mostly bug fixes from now until June!)
 
 #### mobx-35